combustion.rules: Match /module/qemu_fw_cfg instead of the namespace within
During boot, dev-combustion-config.device did not appear even though the fwcfg
key was there. There appears to be some mismatch between what hotplug and
coldplug count as devices, which makes the fwcfg key unusable for this case.
Due to a different bug(?) the platform device triggers the rules but the
systemd tag has no effect. As a workaround, use the module add event as
trigger and check for existence of the key in sysfs. This is reliable enough
as the module registers the sysfs contents synchonously.