Don't consider /var/lib/YaST2/reconfig_system for firstboot detection
This location isn't accessible in the transactional-update environment and
thus the firstboot modules would be missing in initrds on MicroOS in the case
that /etc/machine-id exists but reconfig_system was touched.

So just don't bother. "dracut -f --no-hostonly" and deleting /etc/machine-id
are valid options already to trigger inclusion.